requirements.txt
altair==5.3.0
attrs==23.2.0
bcrypt==4.2.0
blinker==1.8.2
cachetools==5.3.3
captcha==0.6.0
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
duckdb==1.0.0
et-xmlfile==1.1.0
extra-streamlit-components==0.1.71
gitdb==4.0.11
GitPython==3.1.43
idna==3.7
Jinja2==3.1.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
numpy==1.26.4
openpyxl==3.1.3
packaging==24.0
pandas==2.2.2
patsy==0.5.6
pillow==10.3.0
plotly==5.22.0
protobuf==4.25.3
pyarrow==16.1.0
pydeck==0.9.1
Pygments==2.18.0
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rich==13.7.1
rpds-py==0.18.1
scipy==1.14.1
six==1.16.0
smmap==5.0.1
statsmodels==0.14.2
streamlit==1.35.0
streamlit-authenticator==0.3.3
tenacity==8.3.0
toml==0.10.2
toolz==0.12.1
tornado==6.4
typing_extensions==4.12.1
tzdata==2024.1
urllib3==2.2.1
watchdog==4.0.1
